feat(models): expose TrustedRoot, SigningConfig, ClientTrustConfig
Promote TrustedRoot, SigningConfig, and ClientTrustConfig to the
public API via sigstore.models
