+++ sigstore/_internal/trust.py
@@ -19,10 +19,8 @@ Client trust configuration and trust root management for sigstore-python.
 from __future__ import annotations
 
 import logging
-from collections import defaultdict
-from collections.abc import Iterable
 from dataclasses import dataclass
-from datetime import datetime, timezone
+from datetime import datetime
 from enum import Enum
 from pathlib import Path
 from typing import ClassVar, NewType
@@ -38,19 +36,14 @@ from cryptography.x509 import (
 from sigstore_models.common import v1 as common_v1
 from sigstore_models.trustroot import v1 as trustroot_v1
 
-from sigstore._internal.fulcio.client import FulcioClient
-from sigstore._internal.rekor import RekorLogSubmitter
-from sigstore._internal.rekor.client import RekorClient
-from sigstore._internal.rekor.client_v2 import RekorV2Client
-from sigstore._internal.timestamp import TimestampAuthorityClient
-from sigstore._internal.tuf import DEFAULT_TUF_URL, STAGING_TUF_URL, TrustUpdater
 from sigstore._utils import (
     KeyID,
     PublicKey,
+    is_timerange_valid,
     key_id,
     load_der_public_key,
 )
-from sigstore.errors import Error, MetadataError, TUFError, VerificationError
+from sigstore.errors import Error, VerificationError
 
 # Versions supported by this client
 REKOR_VERSIONS = [1, 2]
@@ -59,30 +52,6 @@ FULCIO_VERSIONS = [1]
 OIDC_VERSIONS = [1]
 
 _logger = logging.getLogger(__name__)
-
-
-def _is_timerange_valid(
-    period: common_v1.TimeRange | None, *, allow_expired: bool
-) -> bool:
-    """
-    Given a `period`, checks that the the current time is not before `start`. If
-    `allow_expired` is `False`, also checks that the current time is not after
-    `end`.
-    """
-    now = datetime.now(timezone.utc)
-
-    # If there was no validity period specified, the key is always valid.
-    if not period:
-        return True
-
-    # Active: if the current time is before the starting period, we are not yet
-    # valid.
-    if now < period.start:
-        return False
-
-    # If we want Expired keys, the key is valid at this point. Otherwise, check
-    # that we are within range.
-    return allow_expired or (period.end is None or now <= period.end)
 
 
 @dataclass(init=False)
@@ -294,374 +263,6 @@ class CertificateAuthority:
         The certificates are returned in order from leaf to root, with any
         intermediate certificates in between.
         """
-        if not _is_timerange_valid(self._inner.valid_for, allow_expired=allow_expired):
+        if not is_timerange_valid(self._inner.valid_for, allow_expired=allow_expired):
             return []
         return self._certificates
-
-
-class SigningConfig:
-    """
-    Signing configuration for a Sigstore instance.
-    """
-
-    class SigningConfigType(str, Enum):
-        """
-        Known Sigstore signing config media types.
-        """
-
-        SIGNING_CONFIG_0_2 = "application/vnd.dev.sigstore.signingconfig.v0.2+json"
-
-        def __str__(self) -> str:
-            """Returns the variant's string value."""
-            return self.value
-
-    def __init__(
-        self, inner: trustroot_v1.SigningConfig, tlog_version: int | None = None
-    ):
-        """
-        Construct a new `SigningConfig`.
-
-        tlog_version is an optional argument that enforces that only specified
-        versions of rekor are included in the transparency logs.
-
-        @api private
-        """
-        self._inner = inner
-
-        # must have a recognized media type.
-        try:
-            SigningConfig.SigningConfigType(self._inner.media_type)
-        except ValueError:
-            raise Error(f"unsupported signing config format: {self._inner.media_type}")
-
-        # Create lists of service protos that are valid, selected by the service
-        # configuration & supported by this client
-        if tlog_version is None:
-            tlog_versions = REKOR_VERSIONS
-        else:
-            tlog_versions = [tlog_version]
-
-        self._tlogs = self._get_valid_services(
-            self._inner.rekor_tlog_urls, tlog_versions, self._inner.rekor_tlog_config
-        )
-        if not self._tlogs:
-            raise Error("No valid Rekor transparency log found in signing config")
-
-        self._tsas = self._get_valid_services(
-            self._inner.tsa_urls, TSA_VERSIONS, self._inner.tsa_config
-        )
-
-        self._fulcios = self._get_valid_services(
-            self._inner.ca_urls, FULCIO_VERSIONS, None
-        )
-        if not self._fulcios:
-            raise Error("No valid Fulcio CA found in signing config")
-
-        self._oidcs = self._get_valid_services(
-            self._inner.oidc_urls, OIDC_VERSIONS, None
-        )
-
-    @classmethod
-    def from_file(
-        cls,
-        path: str,
-    ) -> SigningConfig:
-        """Create a new signing config from file"""
-        inner = trustroot_v1.SigningConfig.from_json(Path(path).read_bytes())
-        return cls(inner)
-
-    @staticmethod
-    def _get_valid_services(
-        services: list[trustroot_v1.Service],
-        supported_versions: list[int],
-        config: trustroot_v1.ServiceConfiguration | None,
-    ) -> list[trustroot_v1.Service]:
-        """Return supported services, taking SigningConfig restrictions into account"""
-
-        # split services by operator, only include valid services
-        services_by_operator: dict[str, list[trustroot_v1.Service]] = defaultdict(list)
-        for service in services:
-            if service.major_api_version not in supported_versions:
-                continue
-
-            if not _is_timerange_valid(service.valid_for, allow_expired=False):
-                continue
-
-            services_by_operator[service.operator].append(service)
-
-        # build a list of services but make sure we only include one service per operator
-        # and use the highest version available for that operator
-        result: list[trustroot_v1.Service] = []
-        for op_services in services_by_operator.values():
-            op_services.sort(key=lambda s: s.major_api_version)
-            result.append(op_services[-1])
-
-        # Depending on ServiceSelector, prune the result list
-        if not config or config.selector == trustroot_v1.ServiceSelector.ALL:
-            return result
-
-        # handle EXACT and ANY selectors
-        count = (
-            config.count
-            if config.selector == trustroot_v1.ServiceSelector.EXACT and config.count
-            else 1
-        )
-
-        if (
-            config.selector == trustroot_v1.ServiceSelector.EXACT
-            and len(result) < count
-        ):
-            raise ValueError(
-                f"Expected {count} services in signing config, found {len(result)}"
-            )
-
-        return result[:count]
-
-    def get_tlogs(self) -> list[RekorLogSubmitter]:
-        """
-        Returns the rekor transparency log clients to sign with.
-        """
-        result: list[RekorLogSubmitter] = []
-        for tlog in self._tlogs:
-            if tlog.major_api_version == 1:
-                result.append(RekorClient(tlog.url))
-            elif tlog.major_api_version == 2:
-                result.append(RekorV2Client(tlog.url))
-            else:
-                raise AssertionError(f"Unexpected Rekor v{tlog.major_api_version}")
-        return result
-
-    def get_fulcio(self) -> FulcioClient:
-        """
-        Returns a Fulcio client to get a signing certificate from
-        """
-        return FulcioClient(self._fulcios[0].url)
-
-    def get_oidc_url(self) -> str:
-        """
-        Returns url for the OIDC provider that client should use to interactively
-        authenticate.
-        """
-        if not self._oidcs:
-            raise Error("No valid OIDC provider found in signing config")
-        return self._oidcs[0].url
-
-    def get_tsas(self) -> list[TimestampAuthorityClient]:
-        """
-        Returns timestamp authority clients for urls configured in signing config.
-        """
-        return [TimestampAuthorityClient(s.url) for s in self._tsas]
-
-
-class TrustedRoot:
-    """
-    The cryptographic root(s) of trust for a Sigstore instance.
-    """
-
-    class TrustedRootType(str, Enum):
-        """
-        Known Sigstore trusted root media types.
-        """
-
-        TRUSTED_ROOT_0_1 = "application/vnd.dev.sigstore.trustedroot+json;version=0.1"
-
-        def __str__(self) -> str:
-            """Returns the variant's string value."""
-            return self.value
-
-    def __init__(self, inner: trustroot_v1.TrustedRoot):
-        """
-        Construct a new `TrustedRoot`.
-
-        @api private
-        """
-        self._inner = inner
-        self._verify()
-
-    def _verify(self) -> None:
-        """
-        Performs various feats of heroism to ensure that the trusted root
-        is well-formed.
-        """
-
-        # The trusted root must have a recognized media type.
-        try:
-            TrustedRoot.TrustedRootType(self._inner.media_type)
-        except ValueError:
-            raise Error(f"unsupported trusted root format: {self._inner.media_type}")
-
-    @classmethod
-    def from_file(
-        cls,
-        path: str,
-    ) -> TrustedRoot:
-        """Create a new trust root from file"""
-        inner = trustroot_v1.TrustedRoot.from_json(Path(path).read_bytes())
-        return cls(inner)
-
-    def _get_tlog_keys(
-        self, tlogs: list[trustroot_v1.TransparencyLogInstance], purpose: KeyringPurpose
-    ) -> Iterable[common_v1.PublicKey]:
-        """
-        Yields an iterator of public keys for transparency log instances that
-        are suitable for `purpose`.
-        """
-        allow_expired = purpose is KeyringPurpose.VERIFY
-        for tlog in tlogs:
-            if not _is_timerange_valid(
-                tlog.public_key.valid_for, allow_expired=allow_expired
-            ):
-                continue
-
-            yield tlog.public_key
-
-    def rekor_keyring(self, purpose: KeyringPurpose) -> RekorKeyring:
-        """Return keyring with keys for Rekor."""
-
-        keys: list[common_v1.PublicKey] = list(
-            self._get_tlog_keys(self._inner.tlogs, purpose)
-        )
-        if len(keys) == 0:
-            raise MetadataError("Did not find any Rekor keys in trusted root")
-        return RekorKeyring(Keyring(keys))
-
-    def ct_keyring(self, purpose: KeyringPurpose) -> CTKeyring:
-        """Return keyring with key for CTFE."""
-        ctfes: list[common_v1.PublicKey] = list(
-            self._get_tlog_keys(self._inner.ctlogs, purpose)
-        )
-        if not ctfes:
-            raise MetadataError("CTFE keys not found in trusted root")
-        return CTKeyring(Keyring(ctfes))
-
-    def get_fulcio_certs(self) -> list[Certificate]:
-        """Return the Fulcio certificates."""
-
-        certs: list[Certificate] = []
-
-        # Return expired certificates too: they are expired now but may have
-        # been active when the certificate was used to sign.
-        for authority in self._inner.certificate_authorities:
-            certificate_authority = CertificateAuthority(authority)
-            certs.extend(certificate_authority.certificates(allow_expired=True))
-
-        if not certs:
-            raise MetadataError("Fulcio certificates not found in trusted root")
-        return certs
-
-    def get_timestamp_authorities(self) -> list[CertificateAuthority]:
-        """
-        Return the TSA present in the trusted root.
-
-        This list may be empty and in this case, no timestamp verification can be
-        performed.
-        """
-        certificate_authorities: list[CertificateAuthority] = [
-            CertificateAuthority(cert_chain)
-            for cert_chain in self._inner.timestamp_authorities
-        ]
-        return certificate_authorities
-
-
-class ClientTrustConfig:
-    """
-    Represents a Sigstore client's trust configuration, including a root of trust.
-    """
-
-    class ClientTrustConfigType(str, Enum):
-        """
-        Known Sigstore client trust config media types.
-        """
-
-        CONFIG_0_1 = "application/vnd.dev.sigstore.clienttrustconfig.v0.1+json"
-
-        def __str__(self) -> str:
-            """Returns the variant's string value."""
-            return self.value
-
-    @classmethod
-    def from_json(cls, raw: str) -> ClientTrustConfig:
-        """
-        Deserialize the given client trust config.
-        """
-        inner = trustroot_v1.ClientTrustConfig.from_json(raw)
-        return cls(inner)
-
-    @classmethod
-    def production(
-        cls,
-        offline: bool = False,
-    ) -> ClientTrustConfig:
-        """Create new trust config from Sigstore production TUF repository.
-
-        If `offline`, will use data in local TUF cache. Otherwise will
-        update the data from remote TUF repository.
-        """
-        return cls.from_tuf(DEFAULT_TUF_URL, offline)
-
-    @classmethod
-    def staging(
-        cls,
-        offline: bool = False,
-    ) -> ClientTrustConfig:
-        """Create new trust config from Sigstore staging TUF repository.
-
-        If `offline`, will use data in local TUF cache. Otherwise will
-        update the data from remote TUF repository.
-        """
-        return cls.from_tuf(STAGING_TUF_URL, offline)
-
-    @classmethod
-    def from_tuf(
-        cls,
-        url: str,
-        offline: bool = False,
-    ) -> ClientTrustConfig:
-        """Create a new trust config from a TUF repository.
-
-        If `offline`, will use data in local TUF cache. Otherwise will
-        update the trust config from remote TUF repository.
-        """
-        updater = TrustUpdater(url, offline)
-
-        tr_path = updater.get_trusted_root_path()
-        inner_tr = trustroot_v1.TrustedRoot.from_json(Path(tr_path).read_bytes())
-
-        try:
-            sc_path = updater.get_signing_config_path()
-            inner_sc = trustroot_v1.SigningConfig.from_json(Path(sc_path).read_bytes())
-        except TUFError as e:
-            raise e
-
-        return cls(
-            trustroot_v1.ClientTrustConfig(
-                media_type=ClientTrustConfig.ClientTrustConfigType.CONFIG_0_1.value,
-                trusted_root=inner_tr,
-                signing_config=inner_sc,
-            )
-        )
-
-    def __init__(self, inner: trustroot_v1.ClientTrustConfig) -> None:
-        """
-        @api private
-        """
-        self._inner = inner
-
-        # This can be used to enforce a specific rekor major version in signingconfig
-        self.force_tlog_version: int | None = None
-
-    @property
-    def trusted_root(self) -> TrustedRoot:
-        """
-        Return the interior root of trust, as a `TrustedRoot`.
-        """
-        return TrustedRoot(self._inner.trusted_root)
-
-    @property
-    def signing_config(self) -> SigningConfig:
-        """
-        Return the interior root of trust, as a `SigningConfig`.
-        """
-        return SigningConfig(
-            self._inner.signing_config, tlog_version=self.force_tlog_version
-        )
